+++ docs/types/components/organisms/drawer.d.ts
@@ -1,0 +1,8 @@
+// Types
+import type { Dispatch, SetStateAction } from 'react'
+
+export type DrawerProps = {
+	lng: string
+	open: boolean
+	setOpen: Dispatch<SetStateAction<boolean>>
+}